Backend/Api/requirements.txt
fastapi
uvicorn[standard]
python-socketio[asgi]==5.11.1
python-engineio==4.9.0
sqlalchemy
python-dotenv
pydantic
psycopg2-binary
uuid6
paho-mqtt
requests
fastapi_mail
boto3
flake8
pyjwt
bcrypt
ormar
asyncpg